llm_caller/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-caller"
version = "0.1.0"
description = "Multi-backend LLM client with tool calling and multi-turn loops"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84",
    "google-genai>=1.0",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["llm_caller"]
